requirements.txt
flask
torch
torchvision
sklearn
pillow
pytesseract
matplotlib
prometheus_flask_exporter
